requirements.txt
Appium-Python-Client
selenium
python-dotenv